+++ straxen/contexts.py
@@ -1,30 +1,32 @@
 import warnings
 
-from frozendict import frozendict
+from immutabledict import immutabledict
 import strax
 import straxen
 
 
 common_opts = dict(
-   register_all=[
-       straxen.pulse_processing,
-       straxen.peaklet_processing,
-       straxen.peak_processing,
-       straxen.event_processing,
-       straxen.nveto_recorder,
-       straxen.nveto_pulse_processing],
-   store_run_fields=(
-       'name', 'number',
-       'reader.ini.name', 'tags.name',
-       'start', 'end', 'livetime',
-       'trigger.events_built'),
-   check_available=('raw_records', 'records', 'peaklets',
-                    'events', 'event_info'))
+    register_all=[
+        straxen.pulse_processing,
+        straxen.peaklet_processing,
+        straxen.peak_processing,
+        straxen.event_processing,
+        straxen.double_scatter,
+        straxen.nveto_recorder,
+        straxen.nveto_pulse_processing],
+    store_run_fields=(
+        'name', 'number',
+        'reader.ini.name', 'tags.name',
+        'start', 'end', 'livetime',
+        'trigger.events_built'),
+    check_available=('raw_records', 'records', 'peaklets',
+                     'events', 'event_info'))
+
 
 x1t_common_config = dict(
     check_raw_record_overlaps=False,
     n_tpc_pmts=248,
-    channel_map=frozendict(
+    channel_map=immutabledict(
         # (Minimum channel, maximum channel)
         tpc=(0, 247),
         diagnostic=(248, 253),
@@ -53,12 +55,12 @@ xnt_common_config = dict(
                 0.005),
 #    gain_model_nveto=('to_pe_constant',  #otherwise we get all the time warnings...
 #                      0.005),
-    channel_map=frozendict(
+    channel_map=immutabledict(
          # (Minimum channel, maximum channel)
          # Channels must be listed in a ascending order!
          tpc=(0, 493),
          he=(500, 752),  # high energy
-         aqmon=(799, 807),
+         aqmon=(790, 807),
          aqmonnv=(808, 815),  # nveto acquisition monitor
          tpc_blank=(999, 999),
          mv=(1000, 1083),
@@ -68,6 +70,10 @@ xnt_common_config = dict(
     )
 )
 
+
+##
+# XENONnT
+##
 
 
 def xenonnt_online(output_folder='./strax_data',
@@ -148,6 +154,7 @@ def demo():
             storage=[strax.DataDirectory('./strax_data'),
                      strax.DataDirectory('./strax_test_data', readonly=True)],
             register=straxen.RecordsFromPax,
+            free_options=('channel_map',),
             forbid_creation_of=('raw_records',),
             config=dict(**x1t_common_config),
             **common_opts)
@@ -171,15 +178,6 @@ def fake_daq():
         **common_opts)
 
 
-def strax_workshop_dali():
-    warnings.warn(
-        "The strax_workshop_dali context is deprecated and will "
-        "be removed in April 2020. Please use "
-        "straxen.contexts.xenon1t_dali() instead.",
-        DeprecationWarning)
-    return xenon1t_dali()
-
-
 def xenon1t_dali(output_folder='./strax_data', build_lowlevel=False):
     return strax.Context(
         storage=[
@@ -196,6 +194,7 @@ def xenon1t_dali(output_folder='./strax_data', build_lowlevel=False):
             strax.DataDirectory(output_folder,
                                 provide_run_metadata=False)],
         register=straxen.RecordsFromPax,
+        free_options=('channel_map',),
         config=dict(**x1t_common_config),
         # When asking for runs that don't exist, throw an error rather than
         # starting the pax converter
@@ -211,6 +210,7 @@ def xenon1t_led(**kwargs):
     return st.new_context(
         replace=True,
         register=[straxen.RecordsFromPax, straxen.LEDCalibration],
+        free_options=('channel_map',),
         config=st.config,
         storage=st.storage,
         **st.context_config)