+++ straxen/plugins/__init__.py
@@ -16,6 +16,9 @@ from .peak_processing import *
 from . import event_processing
 from .event_processing import *
 
+from . import double_scatter
+from .double_scatter import *
+
 from . import led_calibration
 from .led_calibration import *